ATM-120 Ethereum coins balance auto update service (#105)
* Added interface for EthereumNotifier service.

* Implemented Start and Stop of EthereumNotifier.

* Added basic implementation for subscribe on balance update.

* Implemented RunBalanceChecker runner for ether.

* Updated IEthereumNotier interface with token methods.

* Added logic with incrementing startBlock from last transaction in result.

* Refactor EthereumNotifier with clean up logs and catching CancelationOperationExceptin.

* Removed websocket connection and updated interface of EthereumNotifier.

* Simplified SubscribeOnBalanceUpdate signature.

* Added EthereumBalanceUpdater service.

* Added handling logic on uncontrolled error in requesting txlist from etherscan.io.

* Implemented logic of getting lastBlockNumber from anonymous api and refactored start and stop methods of EtherscanNotifier.

* Refactor ServiceSubscription of TzktEvents model with long LastState type.

+++ Atomex.TzktEvents/Models/ServiceSubscription.cs
@@ -2,5 +2,5 @@
 
 namespace Atomex.TzktEvents.Models
 {
-    public record ServiceSubscription(Action<string> Handler, int LastState = 0);
+    public record ServiceSubscription(Action<string> Handler, long LastState = 0);
 }
